requirements.txt
python-binance
aiohttp
pyyaml
